serverless/src/requirements.txt
requests==2.31.0
scikit-learn==1.0.2
pandas==2.0.3
google-cloud-storage
numpy==1.21.6